Framework2D - Make Gravity, but Model objects can't collide with ground and each other

+++ FrameWork - 2D/src/DX/DXEngine/DXEngine.h
@@ -63,7 +63,7 @@ class DXENGINE
 		float rotation;
 		float move;
 		float move_temp;
-		int prevTime;
+		float prevTime;
 
 		const char* m_LTVSfileDIR;
 		const char* m_LTPSfileDIR;
